+++ lib/anagram_length_bounds.dart
@@ -1,0 +1,4 @@
+class AnagramLengthBounds {
+  static final int minimumAnagramLength = 1;
+  static final int maximumAnagramLength = 45;
+}