+++ lib/app.dart
@@ -1,6 +1,6 @@
+import 'package:anagrammatic/anagram_length_bounds.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:anagrammatic/constants.dart';
 import 'package:anagrammatic/home.dart';
 import 'package:anagrammatic/options.dart';
 import 'package:anagrammatic/themes.dart';
@@ -25,8 +25,8 @@ class AnagrammaticAppState extends State<AnagrammaticApp> {
     super.initState();
     options = Options(
       theme: darkTheme,
-      anagramLengthLowerBound: minimumAnagramLength,
-      anagramLengthUpperBound: maximumAnagramLength,
+      anagramLengthLowerBound: AnagramLengthBounds.minimumAnagramLength,
+      anagramLengthUpperBound: AnagramLengthBounds.maximumAnagramLength,
       sortType: SortType.getSortName(SortTypeName.alpha),
     );
   }