+++ lib/constants.dart
@@ -1,5 +1,0 @@
-import 'package:anagrammatic/anagram.dart';
-
-// TODO: Split these into own file
-final int minimumAnagramLength = 1;
-final int maximumAnagramLength = 45;